TrackFinding: hide CKF error implementation
