+++ Core/CMakeLists.txt
@@ -50,6 +50,7 @@ add_subdirectory(src/MagneticField)
 add_subdirectory(src/Material)
 add_subdirectory(src/Propagator)
 add_subdirectory(src/Surfaces)
+add_subdirectory(src/TrackFinding)
 add_subdirectory(src/Utilities)
 add_subdirectory(src/Vertexing)
 add_subdirectory(src/Visualization)
